requirements.txt
numpy
pandas
selenium
beautifulsoup4
requests
python-dotenv
colorama